requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pydantic
openai
torch
streamlit
pandas
python-dotenv
silero-vad
torch
torchaudio
gtts
openai-whisper
groq
langchain-groq
langgraph
agents
state
google-api-python-client
google-auth
google-auth-oauthlib
